visionary_v4/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.31.0
accelerate>=0.21.0
rembg>=2.0.50
opencv-python>=4.8.0
pillow>=10.0.0
